requirements.txt
PyQt6>=6.5.0
playwright>=1.36.0
Pillow>=9.0.0
pyinstaller>=5.13.0
pytest==7.4.0
pytest-qt==4.2.0
pytest-playwright==0.4.0
loguru==0.7.0
nest_asyncio==1.5.6
